reorganize objects
