+++ build.gradle
@@ -10,9 +10,8 @@ repositories {
 // In this section you declare the dependencies for your production and test code
 dependencies {
     compile group: 'io.rest-assured', name: 'rest-assured', version: '4.1.1'
-    compile group: 'io.cucumber', name: 'cucumber-jvm', version: '4.7.2', ext: 'pom'
-    compile group: 'io.cucumber', name: 'cucumber-testng', version: '4.7.2'
-
+    compile group: 'io.cucumber', name: 'cucumber-java', version: '4.8.0'
+    compile group: 'io.cucumber', name: 'cucumber-testng', version: '4.8.0'
     compile group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1.1'
     compile group: 'com.jayway.jsonpath', name: 'json-path', version: '2.1.0'
 }
