+++ src/test/java/org/fundacionjala/coding/tdd/DELETEStoryCommentTest.java
@@ -1,16 +1,19 @@
-package org.fundacionjala.coding;
+package org.fundacionjala.coding.tdd;
 
 import io.restassured.response.Response;
+import org.fundacionjala.coding.RequestManager;
+import org.fundacionjala.coding.RequestSpecFactory;
 import org.testng.Assert;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
-public class PUTStoryCommentTest {
+public class DELETEStoryCommentTest {
 
     private String projectId;
     private String storyId;
     private String storyCommentId;
+    private int statusCode;
 
     @BeforeTest
     public void setUp() {
@@ -37,16 +40,15 @@ public class PUTStoryCommentTest {
     }
 
     @Test
-    public void testPUTStoryComment() {
+    public void testDELETEStoryComment() {
         //When
-        String expectedStoryCommentString = "updated comment";
-        Response response = RequestManager.put(RequestSpecFactory.getRequestSpec("pivotal"),
-                String.format("/projects/%s/stories/%s/comments/%s", projectId, storyId, storyCommentId),
-                "{\"text\":\"" + expectedStoryCommentString + "\"}");
+        String expectedStoryCommentString = "story comment";
+        Response response = RequestManager.delete(RequestSpecFactory.getRequestSpec("pivotal"),
+                String.format("/projects/%s/stories/%s/comments/%s", projectId, storyId, storyCommentId));
 
         //Then
-        String actualStoryCommentString = response.jsonPath().getString("text");
-        Assert.assertEquals(actualStoryCommentString, expectedStoryCommentString);
+        statusCode = response.getStatusCode();
+        Assert.assertEquals(statusCode, 204);
     }
 
     @AfterTest
