+++ src/test/java/org/fundacionjala/coding/tdd/DELETEStoryTest.java
@@ -1,14 +1,18 @@
-package org.fundacionjala.coding;
+package org.fundacionjala.coding.tdd;
 
 import io.restassured.response.Response;
+import org.fundacionjala.coding.RequestManager;
+import org.fundacionjala.coding.RequestSpecFactory;
 import org.testng.Assert;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
-public class POSTStoryTest {
+public class DELETEStoryTest {
 
     private String projectId;
+    private String storyId;
+    private int statusCode;
 
     @BeforeTest
     public void setUp() {
@@ -18,19 +22,24 @@ public class POSTStoryTest {
                 "/projects",
                 "{\"name\":\"" + expectedProjectName + "\"}");
         projectId = response.jsonPath().getString("id");
+
+        //And
+        String expectedStoryString = "story string";
+        response = RequestManager.post(RequestSpecFactory.getRequestSpec("pivotal"),
+                String.format("/projects/%s/stories", projectId),
+                "{\"name\":\"" + expectedStoryString + "\"}");
+        storyId = response.jsonPath().getString("id");
     }
 
     @Test
-    public void testPOSTStory() {
+    public void testDELETEStory() {
         //When
-        String expectedStoryString = "story string";
-        Response response = RequestManager.post(RequestSpecFactory.getRequestSpec("pivotal"),
-                String.format("/projects/%s/stories", projectId),
-                "{\"name\":\"" + expectedStoryString + "\"}");
+        Response response = RequestManager.delete(RequestSpecFactory.getRequestSpec("pivotal"),
+                String.format("/projects/%s/stories/%s", projectId, storyId));
 
         //Then
-        String actualStoryString = response.jsonPath().getString("name");
-        Assert.assertEquals(actualStoryString, expectedStoryString);
+        statusCode = response.getStatusCode();
+        Assert.assertEquals(statusCode, 204);
     }
 
     @AfterTest
