+++ src/test/java/org/fundacionjala/coding/tdd/GETStoryCommentTest.java
@@ -1,15 +1,18 @@
-package org.fundacionjala.coding;
+package org.fundacionjala.coding.tdd;
 
 import io.restassured.response.Response;
+import org.fundacionjala.coding.RequestManager;
+import org.fundacionjala.coding.RequestSpecFactory;
 import org.testng.Assert;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
-public class PUTStoryTest {
+public class GETStoryCommentTest {
 
     private String projectId;
     private String storyId;
+    private String storyCommentId;
 
     @BeforeTest
     public void setUp() {
@@ -26,19 +29,25 @@ public class PUTStoryTest {
                 String.format("/projects/%s/stories", projectId),
                 "{\"name\":\"" + expectedStoryString + "\"}");
         storyId = response.jsonPath().getString("id");
+
+        //And
+        String expectedStoryCommentString = "story comment";
+        response = RequestManager.post(RequestSpecFactory.getRequestSpec("pivotal"),
+                String.format("/projects/%s/stories/%s/comments", projectId, storyId),
+                "{\"text\":\"" + expectedStoryCommentString + "\"}");
+        storyCommentId = response.jsonPath().getString("id");
     }
 
     @Test
-    public void testPUTStory() {
+    public void testGETStoryComment() {
         //When
-        String expectedStoryString = "updated story";
-        Response response = RequestManager.put(RequestSpecFactory.getRequestSpec("pivotal"),
-                String.format("/projects/%s/stories/%s", projectId, storyId),
-                "{\"name\":\"" + expectedStoryString + "\"}");
+        String expectedStoryCommentString = "story comment";
+        Response response = RequestManager.get(RequestSpecFactory.getRequestSpec("pivotal"),
+                String.format("/projects/%s/stories/%s/comments/%s", projectId, storyId, storyCommentId));
 
         //Then
-        String actualStoryString = response.jsonPath().getString("name");
-        Assert.assertEquals(actualStoryString, expectedStoryString);
+        String actualStoryCommentString = response.jsonPath().getString("text");
+        Assert.assertEquals(actualStoryCommentString, expectedStoryCommentString);
     }
 
     @AfterTest
